arrumei dois errinhos

+++ exercicio16.py
@@ -4,7 +4,7 @@ import math
 
 # O 'input' para gerar um prompt para o usuário digitar e 'int' para converter o número digitado,
 #  de string para inteiro e ser possível realizar posteriormente um cálculo
-metro_quadrado_area_a_ser_pintada = int(input("Responda qual metro quadrado da área que irá pintar: "))
+metro_quadrado_area_a_ser_pintada = float(input("Responda qual metro quadrado da área que irá pintar: "))
 
 # A variável metro_quadrado_area_a_ser_pintada será dividida por 3, já que 1 litro cobre 3 metros quadrados,
 #  'math.ceil' arredonda o resultado para cima, pois não é possível comprar a lata fracionada
